Rework of simple md to html tree builder script.
Again, this is a really, really basic scripts.  There are much better
and more efficient options to generate doc sites.  I use it for quick
and dirty, simple sites usually internal.  It has been reworked to
analyze the sitemap and order files appropriately. Technically, it
could now provide a previous and next link as well.